pyproject.toml
[tool.poetry]
name = "django-htmx-101"
version = "0.1.0"
description = ""
authors = ["Giuseppe Rosa <giusepper11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-django = "^4.8.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"
pytest-factoryboy = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
line_length = 79
include_trailing_comma = true

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
